+++ packages/eden-watcher/src/entity/ProducerSetChange.ts
@@ -21,7 +21,7 @@ export class ProducerSetChange {
   @Column('integer')
   blockNumber!: number;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   _blockNumber!: bigint;
 
   @Column('varchar')
